requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
streamlit>=1.29.0
streamlit-chat>=0.1.1
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
chromadb>=0.4.22
python-dotenv>=1.0.0
pandas>=2.1.0
openpyxl>=3.1.2
unstructured>=0.10.30
python-magic-bin>=0.4.14
protobuf>=3.20.0
msoffcrypto-tool>=5.0.0
